+++ libs/api/account/data-access/src/lib/api-account-data-access.module.ts
@@ -1,17 +1,11 @@
 import { ApiAppDataAccessModule } from '@kin-kinetic/api/app/data-access'
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
 import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
-import { ApiTransactionDataAccessModule } from '@kin-kinetic/api/transaction/data-access'
 import { Module } from '@nestjs/common'
 import { ApiAccountDataAccessService } from './api-account-data-access.service'
 
 @Module({
-  imports: [
-    ApiAppDataAccessModule,
-    ApiCoreDataAccessModule,
-    ApiKineticDataAccessModule,
-    ApiTransactionDataAccessModule,
-  ],
+  imports: [ApiAppDataAccessModule, ApiCoreDataAccessModule, ApiKineticDataAccessModule],
   providers: [ApiAccountDataAccessService],
   exports: [ApiAccountDataAccessService],
 })
