+++ libs/api/account/data-access/src/lib/api-account-data-access.service.spec.ts
@@ -1,7 +1,6 @@
 import { ApiAppDataAccessModule } from '@kin-kinetic/api/app/data-access'
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
 import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
-import { ApiTransactionDataAccessModule } from '@kin-kinetic/api/transaction/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiAccountDataAccessService } from './api-account-data-access.service'
 
@@ -10,12 +9,7 @@ describe('ApiAccountDataAccessService', () => {
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [
-        ApiAppDataAccessModule,
-        ApiCoreDataAccessModule,
-        ApiKineticDataAccessModule,
-        ApiTransactionDataAccessModule,
-      ],
+      imports: [ApiAppDataAccessModule, ApiCoreDataAccessModule, ApiKineticDataAccessModule],
       providers: [ApiAccountDataAccessService],
     }).compile()
