+++ libs/api/app/data-access/src/lib/api-app-data-access.module.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { ApiWalletDataAccessModule } from '@kin-kinetic/api/wallet/data-access'
 import { Module } from '@nestjs/common'
 import { ApiAppAdminDataAccessService } from './api-app-admin-data-access.service'
@@ -23,6 +23,6 @@ import { ApiAppUserDataAccessService } from './api-app-user-data-access.service'
     ApiAppEnvUserDataAccessService,
     ApiAppUserDataAccessService,
   ],
-  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWalletDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule, ApiWalletDataAccessModule],
 })
 export class ApiAppDataAccessModule {}
