+++ libs/api/app/data-access/src/lib/api-app-data-access.service.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessService } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessService } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticService } from '@kin-kinetic/api/kinetic/data-access'
 import { Injectable, Logger, NotFoundException, OnModuleInit } from '@nestjs/common'
 import { Counter } from '@opentelemetry/api-metrics'
 import { Prisma } from '@prisma/client'
@@ -38,7 +38,7 @@ export class ApiAppDataAccessService implements OnModuleInit {
   private getAppConfigErrorCounter: Counter
   private getAppConfigSuccessCounter: Counter
 
-  constructor(private readonly data: ApiCoreDataAccessService, private readonly solana: ApiSolanaDataAccessService) {}
+  constructor(private readonly data: ApiCoreDataAccessService, private readonly kinetic: ApiKineticService) {}
 
   async onModuleInit() {
     this.getAppConfigErrorCounter = this.data.metrics.getCounter('api_app_get_app_config_error_counter', {
@@ -101,7 +101,7 @@ export class ApiAppDataAccessService implements OnModuleInit {
 
   async getAppHealth(appKey: string): Promise<AppHealth> {
     const isKineticOk = true
-    const solana = await this.solana.getConnection(appKey)
+    const solana = await this.kinetic.getSolanaConnection(appKey)
 
     const isSolanaOk = await solana.healthCheck()
