+++ libs/api/cluster/data-access/src/lib/api-cluster-admin-data-access.service.ts
@@ -1,6 +1,6 @@
 import { ApiCoreDataAccessService } from '@kin-kinetic/api/core/data-access'
 import { getAppKey } from '@kin-kinetic/api/core/util'
-import { ApiSolanaDataAccessService } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticService } from '@kin-kinetic/api/kinetic/data-access'
 import { Keypair } from '@kin-kinetic/keypair'
 import { BadRequestException, Injectable, Logger } from '@nestjs/common'
 import { MintType, Prisma } from '@prisma/client'
@@ -12,7 +12,7 @@ import { ClusterStatus } from './entity/cluster-status.enum'
 @Injectable()
 export class ApiClusterAdminDataAccessService {
   private readonly logger = new Logger(ApiClusterAdminDataAccessService.name)
-  constructor(private readonly data: ApiCoreDataAccessService, private readonly solana: ApiSolanaDataAccessService) {}
+  constructor(private readonly data: ApiCoreDataAccessService, private readonly kinetic: ApiKineticService) {}
 
   async adminCreateCluster(userId: string, data: AdminClusterCreateInput) {
     await this.data.ensureAdminUser(userId)
@@ -78,7 +78,7 @@ export class ApiClusterAdminDataAccessService {
       include: { app: true },
     })
     for (const env of envs) {
-      this.solana.deleteConnection(getAppKey(env.name, env.app.index))
+      this.kinetic.deleteSolanaConnection(getAppKey(env.name, env.app.index))
     }
     return updated
   }
