+++ libs/api/cluster/data-access/src/lib/api-cluster-data-access.module.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { Module } from '@nestjs/common'
 import { ApiClusterAdminDataAccessService } from './api-cluster-admin-data-access.service'
 import { ApiClusterUserDataAccessService } from './api-cluster-user-data-access.service'
@@ -7,6 +7,6 @@ import { ApiClusterUserDataAccessService } from './api-cluster-user-data-access.
 @Module({
   providers: [ApiClusterAdminDataAccessService, ApiClusterUserDataAccessService],
   exports: [ApiClusterAdminDataAccessService, ApiClusterUserDataAccessService],
-  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule],
 })
 export class ApiClusterDataAccessModule {}
