+++ libs/api/kinetic/data-access/src/index.ts
@@ -1,4 +1,8 @@
 export * from './lib/api-kinetic-data-access.module'
 export * from './lib/api-kinetic.service'
+export * from './lib/entities/get-transaction-response.entity'
+export * from './lib/entities/latest-blockhash-response.entity'
+export * from './lib/entities/minimum-rent-exemption-balance-request.dto'
+export * from './lib/entities/minimum-rent-exemption-balance-response.entity'
 export * from './lib/interfaces/process-transaction-options'
 export * from './lib/interfaces/transaction-with-errors'
