+++ libs/api/kinetic/data-access/src/lib/entities/compiled-inner-instruction.entity.ts
