+++ libs/api/kinetic/data-access/src/lib/entities/compiled-instruction.entity.ts
