+++ libs/api/kinetic/data-access/src/lib/entities/confirmed-transaction-meta.entity.ts
