+++ libs/api/kinetic/data-access/src/lib/entities/get-transaction-response.entity.ts
