+++ libs/api/kinetic/data-access/src/lib/entities/latest-blockhash-response.entity.ts
