+++ libs/api/kinetic/data-access/src/lib/entities/minimum-rent-exemption-balance-response.entity.ts
