+++ libs/api/kinetic/data-access/src/lib/entities/rpc-response-and-context.entity.ts
