+++ libs/api/kinetic/data-access/src/lib/entities/transaction-response.entity.ts
