+++ libs/api/transaction/data-access/src/index.ts
@@ -2,14 +2,9 @@ export * from './lib/api-transaction-data-access.module'
 export * from './lib/api-transaction-data-access.service'
 export * from './lib/api-transaction-user-data-access.service'
 export * from './lib/dto/make-transfer-request.dto'
-export * from './lib/dto/minimum-rent-exemption-balance-request.dto'
 export * from './lib/dto/user-transaction-list.input'
-export * from './lib/entities/get-transaction-response.entity'
-export * from './lib/entities/latest-blockhash.entity'
-export * from './lib/entities/minimum-rent-exemption-balance-response.entity'
 export * from './lib/entities/transaction-counter.entity'
 export * from './lib/entities/transaction-error-type.enum'
 export * from './lib/entities/transaction-error.entity'
 export * from './lib/entities/transaction-status.enum'
 export * from './lib/entities/transaction.entity'
-export * from './lib/helpers/parse-error'
