--- libs/api/transaction/data-access/src/lib/helpers/parse-error.ts
@@ -1,21 +1,0 @@
-import { Prisma, TransactionErrorType } from '@prisma/client'
-
-export function parseError(
-  error,
-  errorType?: string,
-  instruction?: number,
-): Prisma.TransactionErrorCreateWithoutTransactionInput {
-  const message = error?.toString()
-  let type: TransactionErrorType = TransactionErrorType.Unknown
-
-  if (errorType === 'InvalidAccountData') {
-    type = TransactionErrorType.InvalidAccount
-  }
-
-  return {
-    logs: error?.logs,
-    message: message || ' ** NO ERROR MESSAGE ** ',
-    instruction,
-    type,
-  }
-}
